added comments as a group and rewrote some things in texthandler

+++ src/main/java/bfst19/Route_parsing/Edge.java
@@ -86,7 +86,7 @@ public class Edge implements Serializable {
         }
     }
 
-    public boolean isForwardAllowed(Vehicle type, long id) {
+    public boolean isForwardAllowed(Vehicle type, int id) {
         Drivabillity drivable = getDrivableFromVehicleType(type);
         if(drivable==Drivabillity.BOTHWAYS){
             return true;
